grpc_servicer/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smg-grpc-servicer"
version = "0.5.2"
description = "SMG gRPC servicer implementations for LLM inference engines (vLLM, SGLang, TokenSpeed)"
requires-python = ">=3.10"
dependencies = [
    "smg-grpc-proto>=0.4.6",
    "grpcio>=1.78.0",
    "grpcio-reflection>=1.78.0",
    "grpcio-health-checking>=1.78.0",
]
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    {name = "Chang Su", email = "mckvtl@gmail.com"},
    {name = "Simo Lin", email = "linsimo.mark@gmail.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
vllm = ["vllm>=0.19.0"]
sglang = ["sglang>=0.5.10"]
tokenspeed = [
    # TokenSpeed is installed out-of-tree (usually as an editable ``pip install -e``
    # against the sibling checkout). We don't hard-pin a PyPI version here so the
    # extra can be used in dev containers that mount the repo directly.
    "uvloop>=0.19.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "tokenspeed: tests that require TokenSpeed",
]

[project.urls]
Homepage = "https://github.com/lightseekorg/smg"
Repository = "https://github.com/lightseekorg/smg"

[tool.setuptools.packages.find]
where = ["."]
include = ["smg_grpc_servicer*"]
